moved status flags into separate file
