+++ src/ml_gym/multiprocessing/states.py
@@ -1,0 +1,12 @@
+from enum import Enum, IntEnum
+
+
+class JobType(IntEnum):
+    CALC = 1
+    TERMINATE = 2
+
+
+class JobStatus(str, Enum):
+    INIT = "INIT"
+    RUNNING = "RUNNING"
+    DONE = "DONE"
